backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.30.0
alembic>=1.13.3
pydantic>=2.9.2
pydantic-settings>=2.6.0
pydantic[email]>=2.9.2
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.12
python-dotenv>=1.0.1
cloudinary>=1.36.0
httpx>=0.27.0
pyzbar>=0.1.9
Pillow>=10.0.0
openai>=1.0.0
pytesseract>=0.3.10
numpy>=1.24.0
